Updated the review comments.
