+++ callautomation-appointment-reminder/config.ini
@@ -4,15 +4,10 @@
 Connectionstring=%Connection string%
 #<!-- Phone number provisioned for the resource (in E.164 Format, e.g. +14251002000). -->
 SourcePhone=%Source Phone%
-#<!-- Destination identities to call. Multiple outbound calls are seperated by a semi-colon and participants in an outbound call are seperated by a coma. 
-#For e.g. +14251002000, 8:acs:ab12b0ea-85ea-4f83-b0b6-84d90209c7c4_00000009-bce0-da09-54b7-xxxxxxxxxxxx; +14251002001, 8:acs:ab12b0ea-85ea-4f83-b0b6-84d90209c7c4_00000009-bce0-da09-555-xxxxxxxxxxxx). -->
 TargetIdentity=%Target Identity%
-
-
 EventCallBackRoute= /api/callbacks
-AppBaseUri= %App Base Uri% 
+App_base_uri= %App Base Uri% 
 AppointmentReminderMenuAudio= /Audio/AppointmentReminderMenu.wav
-# AddParticipantNumber= +918374734062
 AppointmentConfirmedAudio= /audio/AppointmentConfirmedAudio.wav
 AppointmentCancelledAudio= /audio/AppointmentCancelledAudio.wav
 AgentAudio= /audio/AgentAudio.wav
