+++ callautomation-appointment-reminder/requirements.txt
@@ -1,31 +1,3 @@
 aiohttp==3.7.4.post0
-async-timeout==3.0.1
-attrs==21.2.0
-azure-cognitiveservices-speech==1.27.0
-azure-common==1.1.27
-azure-communication-callautomation @file:///C:/pyenv311/azure_communication_callautomation-1.0.0a20230413001-py3-none-any.whl
-azure-communication-chat==1.1.0
+azure-communication-callautomation @file:///C:/pyenv2_311/azure_communication_callautomation-1.0.0a20230413001-py3-none-any.whl
 azure-communication-identity==1.3.1
-azure-core==1.26.4
-azure-nspkg==3.0.2
-azure-storage==0.36.0
-certifi==2021.5.30
-cffi==1.15.1
-chardet==4.0.0
-charset-normalizer==2.0.4
-cryptography==3.4.8
-idna==3.2
-isodate==0.6.0
-msrest==0.7.1
-multidict==6.0.4
-nest-asyncio==1.5.1
-oauthlib==3.1.1
-psutil==5.8.0
-pycparser==2.20
-python-dateutil==2.8.2
-requests==2.26.0
-requests-oauthlib==1.3.0
-six==1.16.0
-typing-extensions==4.3.0
-urllib3==1.26.6
-yarl==1.8.2
